Adjust classes/packages names/locations
