+++ parsers/src/main/java/org/iqu/parsers/entities/NewsArticle.java
@@ -1,8 +1,14 @@
-package entities;
+package org.iqu.parsers.entities;
 
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Holds data representing a news article.
+ * 
+ * @author Cristi Badoi
+ *
+ */
 public class NewsArticle {
   private long date;
   private String id;
@@ -126,6 +132,9 @@ public class NewsArticle {
         + ", images=" + images + ", thumbnail_id=" + thumbnail_id + ", external_url=" + external_url + "]";
   }
 
+  /**
+   * Based solely on id.
+   */
   @Override
   public int hashCode() {
     final int prime = 31;
@@ -134,6 +143,9 @@ public class NewsArticle {
     return result;
   }
 
+  /**
+   * Based solely on id.
+   */
   @Override
   public boolean equals(Object obj) {
     if (this == obj)
